codebase/src/db_service/requirements.txt
psycopg2-binary==2.9.10
grpcio==1.66.2
grpcio-tools==1.66.2
